Renamed turnip monkey patch
